+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Transition.hs
@@ -8,11 +8,14 @@ module Cardano.Ledger.Alonzo.Transition (
   TransitionConfig (..),
 ) where
 
+import Cardano.Ledger.Alonzo.Core (AlonzoEraPParams, ppCostModelsL)
 import Cardano.Ledger.Alonzo.Era
-import Cardano.Ledger.Alonzo.Genesis (AlonzoGenesis)
+import Cardano.Ledger.Alonzo.Genesis
 import Cardano.Ledger.Alonzo.Translation ()
 import Cardano.Ledger.Mary
 import Cardano.Ledger.Mary.Transition (TransitionConfig (MaryTransitionConfig))
+import Cardano.Ledger.Plutus.CostModels (CostModels)
+import Cardano.Ledger.Shelley.LedgerState
 import Cardano.Ledger.Shelley.Transition
 import GHC.Generics
 import Lens.Micro
@@ -27,7 +30,13 @@ instance EraTransition AlonzoEra where
 
   mkTransitionConfig = AlonzoTransitionConfig
 
-  injectIntoTestState = shelleyRegisterInitialFundsThenStaking
+  injectIntoTestState cfg = 
+    shelleyRegisterInitialFundsThenStaking cfg . alonzoInjectCostModels cfg
+
+alonzoInjectCostModels cfg =
+  case agExtraConfig $ cfg ^. tcTranslationContextL of
+    Nothing -> id
+    Just aec -> overrideCostModels (aecCostModels aec)
 
   tcPreviousEraConfigL =
     lens atcMaryTransitionConfig (\atc pc -> atc {atcMaryTransitionConfig = pc})
@@ -36,3 +45,13 @@ instance EraTransition AlonzoEra where
     lens atcAlonzoGenesis (\atc ag -> atc {atcAlonzoGenesis = ag})
 
 instance NoThunks (TransitionConfig AlonzoEra)
+
+overrideCostModels ::
+  (EraTransition era, AlonzoEraPParams era) =>
+  Maybe CostModels ->
+  NewEpochState era ->
+  NewEpochState era
+overrideCostModels Nothing nes = nes
+overrideCostModels = \case 
+  Nothing -> id 
+  Just cms -> nesEsL . curPParamsEpochStateL . ppCostModelsL %~ updateCostModels cms